refactor: create and populate components dir
