+++ src/app/components/index.ts
@@ -1,0 +1,2 @@
+export { Contact } from './Contact';
+export { Project } from './Project';
