+++ src/app/page.css.ts
@@ -1,5 +1,5 @@
 import { style } from '@vanilla-extract/css';
-import { gray, red } from '@radix-ui/colors';
+import { gray } from '@radix-ui/colors';
 
 // clamps are set to min: 320px, max: 723px
 
@@ -54,45 +54,6 @@ export const projects = style({
   flexDirection: 'column',
   gap: 12,
 });
-export const project = style({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: 2,
-
-  textDecoration: 'none',
-});
-export const projectHeader = style({
-  display: 'flex',
-  justifyContent: 'space-between',
-  alignItems: 'start',
-});
-export const projectName = style({
-  fontSize: 'clamp(1.125rem, 0.976rem + 0.746vw, 1.313rem)',
-
-  selectors: {
-    '&::before': {
-      content: '↗ ',
-      verticalAlign: 'text-top',
-      fontSize: 12,
-    },
-    [`${project}:hover &`]: {
-      textDecoration: 'underline',
-    },
-  },
-});
-export const projectYear = style({
-  color: gray.gray11,
-
-  fontSize: 'clamp(.625rem, 0.426rem + .993vw, .875rem)',
-  fontWeight: 400,
-});
-export const projectDescription = style({
-  color: gray.gray11,
-  fontSize: 'clamp(.688rem, 0.54rem + .742vw, .875rem)',
-  fontWeight: 400,
-
-  maxWidth: '66ch',
-});
 
 export const contacts = style({
   display: 'flex',
@@ -100,23 +61,4 @@ export const contacts = style({
 
   maxWidth: 659,
   paddingBottom: 24,
-});
-export const contact = style({
-  textDecoration: 'none',
-  fontSize: 'clamp(1.125rem, 0.976rem + 0.746vw, 1.313rem)',
-
-  ':hover': {
-    textDecoration: 'underline',
-  },
-  ':visited': {
-    color: red.red11,
-  },
-
-  selectors: {
-    '&::before': {
-      content: '↗ ',
-      verticalAlign: 'text-top',
-      fontSize: 12,
-    },
-  },
 });
