+++ src/app/page.tsx
@@ -1,41 +1,10 @@
 import React, { FC } from 'react';
-
-import Link from 'next/link';
-import * as styles from './page.css';
 import { contacts, projects, bio } from './data';
+import { Project } from './components/Project';
+import { Contact } from './components/Contact';
+import * as styles from './page.css';
 
-interface ProjectProps {
-  name: string;
-  url: string;
-  year: number;
-  description: string;
-}
-
-interface ContactProps {
-  title: string;
-  url: string;
-}
-
-const Project: FC<ProjectProps> = ({ name, url, year, description }) => (
-  <Link className={styles.project} href={url}>
-    <header className={styles.projectHeader}>
-      <span className={styles.projectName}>{name}</span>
-      <span className={styles.projectYear}>{year}</span>
-    </header>
-    <p className={styles.projectDescription}>{description}</p>
-  </Link>
-);
-
-const Contact: FC<ContactProps> = ({ title, url }) => (
-  <a
-    className={styles.contact}
-    href={title === 'Email' ? `mailto:${url}` : url}
-  >
-    {title}
-  </a>
-);
-
-const Index = () => (
+const Index: FC = () => (
   <main className={styles.main}>
     <header>
       <h1 className={styles.mainName}>Gregory Nathan Jinsoo Monroe</h1>
